Remove duplicated code across wasm wrappers
