+++ yarn-project/barretenberg.js/src/wasm/index.ts
@@ -1,1 +1,2 @@
 export { BarretenbergWasm } from './barretenberg_wasm.js';
+export * from './load_crs.js';
