+++ yarn-project/circuits.js/package.json
@@ -39,6 +39,7 @@
     "rootDir": "./src"
   },
   "dependencies": {
+    "@aztec/barretenberg.js": "workspace:^",
     "@aztec/foundation": "workspace:^",
     "@types/lodash.times": "^4.3.7",
     "cross-fetch": "^3.1.5",
