+++ yarn-project/foundation/src/wasm/index.ts
@@ -1,7 +1,6 @@
 export { WebDataStore } from './worker/browser/index.js';
 export { NodeDataStore } from './worker/node/index.js';
 export { WasmWorker, WorkerPool } from './worker/index.js';
-export { WasmModule } from './wasm/index.js';
-export { AsyncCallState, AsyncFnState } from './wasm/index.js';
+export { WasmModule, WasmWrapper, AsyncWasmWrapper, AsyncCallState, AsyncFnState } from './wasm/index.js';
 export { DispatchMsg, WorkerListener, TransportServer, NodeListener } from './transport/index.js';
 export { Transfer, isTransferDescriptor } from './transport/interface/transferable.js';
