+++ yarn-project/foundation/src/wasm/wasm/index.ts
@@ -1,2 +1,4 @@
 export { AsyncCallState, AsyncFnState } from './async_call_state.js';
 export { WasmModule } from './wasm_module.js';
+export { WasmWrapper } from './wasm_wrapper.js';
+export { AsyncWasmWrapper } from './async_wasm_wrapper.js';
