+++ yarn-project/foundation/src/wasm/wasm/wasm_wrapper.ts
@@ -1,0 +1,103 @@
+import { AsyncCallState, AsyncFnState, NodeDataStore, WasmModule, WebDataStore } from '@aztec/foundation/wasm';
+
+import isNode from 'detect-node';
+import { readFile } from 'fs/promises';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+/**
+ * Get the WASM binary.
+ * @param relativePath - Relative path to the code (eg '/barretenberg.wasm')
+ * @returns The binary buffer.
+ */
+export async function fetchCode(relativePath: string) {
+  if (isNode) {
+    const __dirname = dirname(fileURLToPath(import.meta.url));
+    return await readFile(__dirname + relativePath);
+  } else {
+    const res = await fetch(relativePath);
+    return Buffer.from(await res.arrayBuffer());
+  }
+}
+
+/**
+ * A low-level wrapper for an instance of a WASM.
+ */
+export abstract class WasmWrapper {
+  protected store = isNode ? new NodeDataStore() : new WebDataStore();
+  protected wasm!: WasmModule;
+
+  abstract codeRelativePath: string;
+
+  constructor(private loggerName?: string) {}
+
+  /**
+   * 20 pages by default. 20*2**16 \> 1mb stack size plus other overheads.
+   * 8192 maximum by default. 512mb.
+   * @param initial - Initial memory pages.
+   * @param maximum - Max memory pages.
+   * @returns Self
+   */
+  public async init(initial = 20, maximum = 8192): Promise<WasmWrapper> {
+    let wasm: WasmModule;
+    this.wasm = wasm = new WasmModule(
+      await fetchCode(this.codeRelativePath),
+      module => ({
+        /**
+         * Log a string from wasm.
+         * @param addr - The string address to log.
+         */
+        logstr(addr: number) {
+          const str = wasm.getMemoryAsString(addr);
+          const m = wasm.getMemory();
+          const str2 = `${str} (mem: ${(m.length / (1024 * 1024)).toFixed(2)}MB)`;
+          wasm.getLogger()(str2);
+        },
+        memory: module.getRawMemory(),
+        ...this.getImportFns(module),
+      }),
+      this.loggerName,
+    );
+    await wasm.init(initial, maximum);
+    return this;
+  }
+
+  /**
+   * These are functions implementations for imports we've defined are needed.
+   * The native C++ build defines these in a module called "env". We must implement TypeScript versions here.
+   * @param module - The wasm module.
+   * @returns An object of functions called from cpp that need to be answered by ts.
+   */
+  protected getImportFns(module: WasmModule): any {
+    return {};
+  }
+
+  /**
+   * Get a slice of memory between two addresses.
+   * @param start - The start address.
+   * @param end - The end address.
+   * @returns A Uint8Array view of memory.
+   */
+  public getMemorySlice(start: number, end: number) {
+    return this.wasm.getMemorySlice(start, end);
+  }
+
+  /**
+   * Write data into the heap.
+   * @param arr - The data to write.
+   * @param offset - The address to write data at.
+   */
+  public writeMemory(offset: number, arr: Uint8Array) {
+    this.wasm.writeMemory(offset, arr);
+  }
+
+  /**
+   * Calls into the WebAssembly.
+   * @param name - The method name.
+   * @param args - The arguments to the method.
+   * @returns The numeric integer or address result.
+   */
+  public call(name: string, ...args: any): number {
+    return this.wasm.call(name, ...args);
+  }
+}
